+++ hs-src/Language/Scheme/Libraries.hs
@@ -42,6 +42,7 @@ findModuleFile [String file]
       file == "scheme/r5rs/read.sld" ||
       file == "scheme/r5rs/write.sld" ||
       file == "scheme/base.sld" ||
+      file == "husk/pretty-print.sld" ||
 -- TODO: scheme case-lambda (r7rs)
 -- TODO: scheme process-context (r7rs)
 -- TODO: scheme repl (r7rs)
@@ -69,9 +70,8 @@ moduleImport to from (DottedList [Atom iRenamed] (Atom iOrig) : is) = do
   moduleImport to from is
 moduleImport to from [] = do
   return $ LispEnv to
--- DEBUG:
--- moduleImport to from unknown = do
---   (trace ("MODULE IMPORT DEBUG: " ++ show unknown) return) $ Nil ""
+moduleImport _ _ err = do
+  throwError $ Default $ "Unexpected argument to moduleImport: " ++ show err
 
 -- |Copy a binding from one env to another
 divertBinding