+++ tests/libs/raw.scm
@@ -1,0 +1,2 @@
+(define (raw)
+    'a-raw-included-file)