+++ ChangeLog.markdown
@@ -1,4 +1,15 @@
 
+v3.11
+--------
+
+The major change in this release is support for R<sup>7</sup>RS style library syntax in the compiler. This enables functionality that was previously only available in the interpreter, and sets the stage for husk to begin adding support for R<sup>7</sup>RS.
+
+Bug fixes:
+
+- Allow the husk compiler to reference variables that are defined later in the program. For example, the following now compiles: `(define (foo) (bar))` `(define (bar) (foo))`
+- Fixed `string-set!` to allow an expression to be passed as the character argument.
+- Fixed `string-ref` to work when the string argument is passed using a variable.
+
 v3.10
 --------